fixed issue for account creation

+++ app/controllers/registrations_controller.rb
@@ -49,13 +49,14 @@ class RegistrationsController < Devise::RegistrationsController
     else
       existing_user = User.where_case_insensitive('email', sign_up_params[:email]).first
       if existing_user.present?
-        if existing_user.accept_terms?
-          redirect_to after_sign_up_error_path_for(resource), alert: _('That email address is already registered.')
-          return
-        else
+        if existing_user.invitation_token.present? && 
+             !existing_user.accept_terms?
           existing_user.destroy # Destroys the existing user since the accept terms are nil/false.
           # Note any existing role for that user will be deleted too. Added to accommodate issue at:
           # https://github.com/DMPRoadmap/roadmap/issues/322
+        else
+          redirect_to after_sign_up_error_path_for(resource), alert: _('That email address is already registered.')
+          return
         end
       end
